Transfer doc, also added a config for paths, fixed bug in directivity detection
