+++ scenarios/input_output.py
@@ -20,74 +20,6 @@ from scenarios.utils import get_hypo
 from scenarios.utils import rake_to_type
 
 
-def read_event_xml(file):
-    """
-    Read event.xml. 
-
-    Args:
-        file (str): Path to event.xml file.
-
-    Returns:
-        dict: Dictionary with event info.
-
-    """
-    eventtree = ET.parse(file)
-    eventroot = eventtree.getroot()
-    for eq in eventroot.iter('earthquake'):
-        id_str = eq.attrib['id']
-        magnitude = float(eq.attrib['mag'])
-        hlat = float(eq.attrib['lat'])
-        hlon = float(eq.attrib['lon'])
-        hdepth = float(eq.attrib['depth'])
-        if 'rake' in eq.attrib.keys():
-            rake = float(eq.attrib['rake'])
-        else:
-            rake = None
-        lstring = eq.attrib['locstring']
-        if 'description' in eq.attrib.keys():
-            description = eq.attrib['description']
-        else:
-            description = ""
-        if 'type' in eq.attrib.keys():
-            mech = eq.attrib['type']
-        else:
-            mech = "ALL"
-        year = int(eq.attrib['year'])
-        month = int(eq.attrib['month'])
-        day = int(eq.attrib['day'])
-        hour = int(eq.attrib['hour'])
-        minute = int(eq.attrib['minute'])
-        second = int(eq.attrib['second'])
-        if 'directivity' in eq.attrib.keys():
-            directivity = ast.literal_eval(eq.attrib['directivity'])
-        else:
-            directivity = False
-        if 'eventsourcecode' in eq.attrib.keys():
-            eventsourcecode = eq.attrib['eventsourcecode']
-        else:
-            eventsourcecode = None
-
-    sdt = ShakeDateTime(year, month, day, hour, minute, second, int(0))
-
-    event = {'lat': hlat,
-             'lon': hlon,
-             'depth': hdepth,
-             'mag': magnitude,
-             'rake': rake,
-             'id': id_str,
-             'locstring': lstring,
-             'type': mech,
-             'mech': mech,
-             'time': sdt.strftime('%Y-%m-%dT%H:%M:%SZ'),
-             'timezone': 'UTC',
-             'directivity':directivity,
-             'description':description,
-             'eventsourcecode':eventsourcecode,
-             'sdt':sdt}
-
-    return event
-
-
 
 def write_event_xml(input_dir, rdict, directivity):
     """
